clt-chatbot-system/01.workspaces/apps/chatbot-backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
PyJWT==2.9.0
SQLAlchemy==2.0.36
psycopg[binary]==3.2.1
requests==2.32.3
